% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_travel_time.R
\name{calculate_travel_time}
\alias{calculate_travel_time}
\title{Calculate travel time}
\usage{
calculate_travel_time(
  friction_surface,
  points,
  file_name = "outputs/travel_time.tif",
  overwrite_raster = FALSE
)
}
\arguments{
\item{overwrite_raster}{}
}
\description{
Implements Weiss et al method to calculate travel time given a
}
