% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_from_spatraster.R
\name{ext_from_spatraster}
\alias{ext_from_spatraster}
\title{Extent from \code{SpatRaster}}
\usage{
ext_from_spatraster(r)
}
\arguments{
\item{r}{\code{SpatRaster}}
}
\value{
2x2 \code{matrix}
}
\description{
Formats spatial extent for use in \code{get_friction_surface}.
}
\examples{

library(terra)
r <- terra::rast(
    extent = terra::ext(c(111, 112, 0, 1))
  )

ext_from_spatraster(r)
}
