% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_vect_to_matrix.R
\name{ext_vect_to_matrix}
\alias{ext_vect_to_matrix}
\title{Extent vector to matrix}
\usage{
ext_vect_to_matrix(x)
}
\arguments{
\item{x}{\code{numeric} length 4, consisting of \code{c(xmin, xmax, ymin, ymax)}
dimensions of extent}
}
\value{
2x2 \code{matrix}
}
\description{
Extent vector to matrix
}
\examples{

ext_vect_to_matrix(c(111,112,0, 1))

}
