% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friction_surface.R
\name{get_friction_surface}
\alias{get_friction_surface}
\title{Get friction surface}
\usage{
get_friction_surface(
  surface = c("motor2020", "walk2020"),
  file_name = "friction_surface.tif",
  overwrite = FALSE,
  extent = NULL
)
}
\arguments{
\item{surface}{\code{"motor2020"} or \verb{"walk2020}}

\item{file_name}{Character of file name to write raster to disc}

\item{overwrite}{Overwrite \code{file_name} if exists}

\item{extent}{Extent to pass to \code{malariaAtlas::getRaster}: 2x2 matrix
specifying the spatial extent within which raster data is desired, as
returned by sf::st_bbox() - the first column has the minimum, the second
the maximum values; rows 1 & 2 represent the x & y dimensions respectively
(matrix(c("xmin", "ymin","xmax", "ymax"), nrow = 2, ncol = 2, dimnames =
list(c("x", "y"), c("min", "max")))) (use either shp OR extent; if neither
is specified global raster is returned).}
}
\description{
Wrapper function to download friction surfaces via
\code{malariaAtlas::getRaster}
}
\details{
Additional details...
}
\examples{

ext <- matrix(
  data = c("111", "0", "112", 1),
  nrow = 2,
  ncol = 2,
  dimnames = list(
    c("x", "y"),
    c("min", "max")
   )
 )

 get_friction_surface(
   surface = "motor2020",
   file_name = tempfile(fileext = ".tif"),
   extent = ext
 )

}
