% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traveltime.R
\name{get_traveltime}
\alias{get_traveltime}
\title{API Wraper for the Traveltime API}
\usage{
get_traveltime(..., as.spatial = FALSE)
}
\arguments{
\item{appId}{Traveltime-Plattform Application ID}

\item{apiKey}{Traveltime-Plattform API-Key}

\item{location}{vector of lat / long coordinates, example: c(47.233,8.234)}
}
\value{
the results from the search
}
\description{
gets isochrones for traveltimemaps from Traveltime API
}
\examples{
traveltime30 <- get_traveltime(appId="XXXXXXX",
apiKey="XXXXXXXX",
location=c(47.378610,8.54000),
traveltime=1800,
type="public_transport",
departure="2018-10-05T08:00:00Z")
plot(traveltime30)

}
