% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Singapore MRT and LRT data}
\format{
\subsection{\code{stations} A data frame with 563 rows and 2 columns:}{

\describe{
\item{x}{longitude}
\item{y}{latitude}
}
}
}
\source{
Land Transport Authority. (2019). LTA MRT Station Exit (GEOJSON)
(2024) Dataset. data.gov.sg. Retrieved December 10, 2024 from
\url{https://data.gov.sg/datasets/d_b39d3a0871985372d7e1637193335da5/view}
}
\usage{
stations
}
\description{
Locations of all MRT and LRT station exits in Singapore.
}
\keyword{datasets}
