% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime-package.R
\docType{package}
\name{traveltime-package}
\alias{traveltime}
\alias{traveltime-package}
\title{traveltime: Calculate Travel Times Over Space}
\description{
Calculate travel time over a friction surface from a specified set of locations.
}
\author{
\strong{Maintainer}: Gerry Ryan \email{gerard.ryan@telethonkids.org.au} (\href{https://orcid.org/0000-0003-0183-7630}{ORCID})

Authors:
\itemize{
  \item Daniel Weiss
}

}
\keyword{internal}
