% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime_filter.R
\name{traveltime_filter}
\alias{traveltime_filter}
\title{Calculate traveltime between locations}
\usage{
traveltime_filter(appId = "yourAppId", apiKey = "yourApiKey",
  from = NULL, to = NULL, traveltime = 14400, type = NULL,
  departure = NULL)
}
\arguments{
\item{appId}{Traveltime-Plattform Application ID}

\item{apiKey}{Traveltime-Plattform API-Key}

\item{from}{vector of lat / long coordinates, coordinate system WGS84 / espg 4326, example: c(47.233,8.234).}

\item{to}{vector or list of lat / long coordinates, coordinate system WGS84 / espg 4326, example: c(47.233,8.234).}

\item{traveltime}{traveltime in seconds}

\item{type}{transportation mode}

\item{departure}{time of departure, date in extended ISO-8601 format (incl. timezone), example: "2018-08-05T08:00:00Z"}

\item{arrival}{time of arrival, date in extended ISO-8601 format (incl. timezone), example: "2018-08-05T08:00:00Z"}
}
\value{
the results from the search
}
\description{
\code{traveltime_filter} Calculate traveltime between locations. Returns ....
}
\details{
traveltime_filter - retrieve traveltimes for a pair of locations

To call the API via the package you need to get an api-key here \url{http://docs.traveltimeplatform.com/overview/getting-keys/}.
         If you're not using the API for commercial use users can get 10,000 free API queries a month. The following transport modes are supported (type):"cycling"", "cycling_ferry", "driving", "driving+train", "driving_ferry", "public_transport", "walking", "walking+coach", "walking_bus", "walking_ferry" or "walking_train".
}
\examples{
 \donttest{
 #list of destinations
locations <-list(c(47.378610,8.54000),c(47.578620,8.54000))

traveltimes <- traveltime_filter(appId="YourAppId",
                            apiKey="YourApiKey",
                            from = c(47.578610,8.54000),
                            to = locations ,
                            traveltime=2000,
                            departure="2018-12-06T08:00:00Z")

#plot the isochrones
glimpse(traveltimes)
}

}
\references{
\href{http://docs.traveltimeplatform.com/overview/introduction}{Traveltime Plattform API Docs}
}
