% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime_request_route.R
\name{traveltime_filter_req}
\alias{traveltime_filter_req}
\title{#' API Wrapper for the Traveltime API - helper function
#'
#' traveltime_request calls the api and converts the response to an sf object
#'
#' @importFrom httr POST
#' @noRd}
\usage{
traveltime_filter_req(appId = "yourAppId", apiKey = "yourApiKey",
  from = NULL, to = NULL, traveltime = NULL, type = NULL,
  departure = NULL)
}
\description{
#' API Wrapper for the Traveltime API - helper function
#'
#' traveltime_request calls the api and converts the response to an sf object
#'
#' @importFrom httr POST
#' @noRd
}
