% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime_geocode.R
\name{traveltime_geocode}
\alias{traveltime_geocode}
\title{Geocode addresses using the traveltime API}
\usage{
traveltime_geocode(appId = "yourAppId", apiKey = "yourApiKey", address)
}
\arguments{
\item{appId}{Traveltime-Plattform Application ID}

\item{apiKey}{Traveltime-Plattform API-Key}

\item{address}{A character vector of length one containing the address query.}
}
\description{
TODO: enable vectors of addresses to be passed through directly, along with time limit
TODO: enable reverse geocoding
TODO: refine time limit
}
\examples{
\dontrun{
whiteHouse <- c("1600 Pennsylvania Avenue NW, Washington, DC, 20500, United States", "The White House")
# returns the same address
traveltime_geocode(whiteHouse[1])
traveltime_geocode(whitehouse[2])

# loop over a list of addresses
lapply(whiteHouse, traveltime_geocode)

}
}
