% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltime_request.R
\name{traveltime_request}
\alias{traveltime_request}
\title{API Wrapper for the Traveltime API - helper function}
\usage{
traveltime_request(appId, apiKey, location, traveltime, type, departure)
}
\description{
traveltime_request calls the api and converts the response to an sf object
}
