% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_matrix.R
\name{ext_matrix}
\alias{ext_matrix}
\title{Convert extent to matrix}
\usage{
ext_matrix(x, ...)
}
\arguments{
\item{x}{\code{numeric} length 4, consisting of \code{c(xmin, xmax, ymin, ymax)}
dimensions of extent}
}
\value{
2x2 \code{matrix} - explain esoteric format of matrix
}
\description{
This function converts from formats ...
TODO - explain why we need to do this, which is that we need a special esoteric format
}
\examples{
ext_matrix(c(111,112,0, 1))
# TODO
# add examples of all SpatRaster, SpatVector, etc.
}
