% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friction_surface.R
\name{get_friction_surface}
\alias{get_friction_surface}
\title{Get friction surface}
\usage{
get_friction_surface(
  surface = c("motor2020", "walk2020"),
  filename = NULL,
  overwrite = FALSE,
  extent = NULL
)
}
\arguments{
\item{surface}{\code{"motor2020"} or \verb{"walk2020}.}

\item{filename}{\code{character}. File name for output layer.}

\item{overwrite}{Overwrite \code{filename} if exists}

\item{extent}{Spatial extent as one of:
\itemize{
\item a numeric vector specifying \code{c(xmin,   xmax, ymin, ymax)},
\item a \code{terra} \code{SpatExtent}, \code{SpatVector}, or \code{SpatRaster} from which
the extent will be taken, or
\item or 2x2 \code{matrix}
}

See details}
}
\value{
\code{SpatRaster}
}
\description{
Wrapper function to download friction surfaces via
\code{\link[malariaAtlas:getRaster]{malariaAtlas::getRaster()}}.
}
\details{
Convenience wrapper to \code{\link[malariaAtlas:getRaster]{malariaAtlas::getRaster()}} to access
motorised and walking travel friction layers per Weiss et al. 2020, that
adds safety to check existing files before download. Surfaces can be
downloaded directly from:
\url{https://malariaatlas.org/project-resources/accessibility-to-healthcare/}.

Here ---
\itemize{
\item \code{surface = "motor2020"} will download
\code{"Explorer__2020_motorized_friction_surface"}, and
\item \code{surface = "walk2020"} will download
\code{"Explorer__2020_walking_only_friction_surface"}.
}

D. J. Weiss, A. Nelson, C. A. Vargas-Ruiz, K. Gligoric, S., Bavadekar, E.
Gabrilovich, A. Bertozzi-Villa, J. Rozier, H. S. Gibson, T., Shekel, C.
Kamath, A. Lieber, K. Schulman, Y. Shao, V. Qarkaxhija, A. K. Nandi, S. H.
Keddie, S. Rumisha, P. Amratia, R. Arambepola, E. G. Chestnutt, J. J.
Millar, T. L. Symons, E. Cameron, K. E. Battle, S. Bhatt, and P. W.
Gething. Global maps of travel time to healthcare facilities. (2020) Nature
Medicine. \url{https://doi.org/10.1038/s41591-020-1059-1}

\code{extent} is formatted and then passed through to
\code{\link[malariaAtlas:getRaster]{malariaAtlas::getRaster()}} as a 2x2 matrix. It is converted into a matrix
using \code{ext_matrix}. The \code{matrix} format used is as returned by
\code{sf::st_bbox()} --- the first column has the minimum, the second the
maximum values; rows 1 & 2 represent the x & y dimensions respectively:
\code{matrix(c("xmin", "ymin", "xmax", "ymax"), nrow = 2, ncol = 2, dimnames = list(c("x", "y"), c("min", "max")))}. \code{NULL} extent downloads (large)
global layer.

Troubleshooting: if you get a warning \verb{Failed to connect to MAP geoserver},
this is an issue with fetching the friction surface using
\code{malariaAtlas::getRaster}, and not an issue with \code{traveltime} itself. The
server may be down or there may be an issue with that package.
}
\examples{

# for more examples of passing exten types see ?ext_matrix

ext <- c(111, 0, 112, 1)

 get_friction_surface(
   surface = "motor2020",
   extent = ext
 )

}
