% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friction_surface.R
\name{get_friction_surface}
\alias{get_friction_surface}
\title{Get friction surface}
\usage{
get_friction_surface(
  surface = c("motor2020", "walk2020"),
  file_name = NULL,
  overwrite = FALSE,
  extent = NULL
)
}
\arguments{
\item{surface}{\code{"motor2020"} or \verb{"walk2020}.}

\item{file_name}{\code{character}. File name for output layer.}

\item{overwrite}{Overwrite \code{file_name} if exists}

\item{extent}{Spatial extent as either numeric vector specifying \code{c(xmin, xmax, ymin, ymax)}, \code{SpatExtent}, \code{SpatVector} or \code{SpatRaster} (from which
the extent will be taken), or 2x2 \code{matrix} (see details).}
}
\value{
\code{SpatRaster}
}
\description{
Wrapper function to download friction surfaces via
\code{malariaAtlas::getRaster}
}
\details{
Convenience wrapper to \code{malariaAtlas::getRaster} to access motorised
and walking travel friction layers per Weiss et al. 2020, that adds safety to
check existing files before download. Surfaces can be downloaded directly
from:
\url{https://malariaatlas.org/project-resources/accessibility-to-healthcare/}

\code{surface = "motor2020"} downloads
\code{"Explorer__2020_motorized_friction_surface"}.

\code{surface = "walk2020"} downloads
\code{"Explorer__2020_walking_only_friction_surface"}.

D. J. Weiss, A. Nelson, C. A. Vargas-Ruiz, K. Gligoric, S., Bavadekar, E.
Gabrilovich, A. Bertozzi-Villa, J. Rozier, H. S. Gibson, T., Shekel, C.
Kamath, A. Lieber, K. Schulman, Y. Shao, V. Qarkaxhija, A. K. Nandi, S. H.
Keddie, S. Rumisha, P. Amratia, R. Arambepola, E. G. Chestnutt, J. J. Millar,
T. L. Symons, E. Cameron, K. E. Battle, S. Bhatt, and P. W. Gething. Global
maps of travel time to healthcare facilities. (2020) Nature Medicine.
\url{https://doi.org/10.1038/s41591-020-1059-1}

\code{extent} is passed through is to pass to \code{malariaAtlas::getRaster} as a 2x2
matrix. If passed in as a numeric vector, \code{SpatExtent}, \code{SpatVector}, or
\code{SpatRaster}, it is converted into a matrix using \code{ext_vect_to_matrix} and
\code{ext_from_terra}. \code{matrix} format is  as returned by sf::st_bbox() - the
first column has the minimum, the second the maximum values; rows 1 & 2
represent the x & y dimensions respectively (\code{matrix(c("xmin", "ymin","xmax", "ymax"), nrow = 2, ncol = 2, dimnames = list(c("x", "y"), c("min", "max")))})
(use either shp OR extent; if neither is specified global raster is
returned). \code{NULL} extent downloads (large) global layer.

Additional details...
}
\examples{

ext <- matrix(
  data = c(111, 0, 112, 1),
  nrow = 2,
  ncol = 2,
  dimnames = list(
    c("x", "y"),
    c("min", "max")
   )
 )

 get_friction_surface(
   surface = "motor2020",
   extent = ext
 )

}
